fix reifed generics when subclassing a partially reified generic class
